editcourse.html work
Changed the editcourse.html file to reflect the database properly. Added in a findCourse(courseID) method to CourseManagement.py. Took the findUser() method from UserManagement.py and changed it to reflect finding a course instead of a user. Updated the editCourse(View) method in views.py. Currently running into an issues when I try to edit a course on the webpage. The error is "ValueError at /editcourse/
The view TAScheduler.views.view didn't return an HttpResponse object. It returned None instead.".  Not entirely sure what's going on.

Co-Authored-By: Broderick Hebert <70716012+BMHebert@users.noreply.github.com>
Co-Authored-By: YazanSalem <77870522+YazanSalem@users.noreply.github.com>